+++ bridgebot/bridgehand.py
